Collapses separate PlayerActivity classes into a single GameActivity
