+++ app/src/main/AndroidManifest.xml
@@ -12,23 +12,6 @@
         <activity android:name=".MainActivity">
 
         </activity>
-        <activity android:name=".TwoPlayerActivity"
-                  android:screenOrientation="portrait">
-            <!--<intent-filter>-->
-            <!--<action android:name="android.intent.action.MAIN" />-->
-
-            <!--<category android:name="android.intent.category.LAUNCHER" />-->
-            <!--</intent-filter>-->
-        </activity>
-        <activity android:name=".FourPlayerActivity"
-            android:screenOrientation="portrait">
-            <!--<intent-filter>-->
-                <!--<action android:name="android.intent.action.MAIN" />-->
-
-                <!--<category android:name="android.intent.category.LAUNCHER" />-->
-            <!--</intent-filter>-->
-
-        </activity>
         <activity android:name=".PlayerSelectActivity"
                   android:screenOrientation="portrait">
             <intent-filter>
@@ -61,14 +44,8 @@
             <!--<category android:name="android.intent.category.LAUNCHER" />-->
             <!--</intent-filter>-->
         </activity>
-        <activity android:name=".ThreePlayerActivity"
-                  android:screenOrientation="portrait">
-            <!--<intent-filter>-->
-                <!--<action android:name="android.intent.action.MAIN" />-->
-
-                <!--<category android:name="android.intent.category.LAUNCHER" />-->
-            <!--</intent-filter>-->
-        </activity>
+        <activity android:name=".GameActivity"
+                  android:screenOrientation="portrait"/>
 
     </application>
