+++ app/src/main/java/com/team2/mtglifetracker/PlayerSelectActivity.java
@@ -40,7 +40,7 @@ public class PlayerSelectActivity extends AppCompatActivity {
                 // v.getContext() tip comes from:
                 // https://stackoverflow.com/questions/20241857/android-intent-cannot-resolve-constructor
                 Intent layoutSelectIntent = new Intent(v.getContext(), LayoutSelectActivity.class );
-                layoutSelectIntent.putExtra("playerSelectCount", 2);
+                layoutSelectIntent.putExtra("playerCount", 2);
                 startActivity(layoutSelectIntent);
 
             }
@@ -52,7 +52,8 @@ public class PlayerSelectActivity extends AppCompatActivity {
                 // on click, send user to two player layout selection screen and store choice
                 // v.getContext() tip comes from:
                 // https://stackoverflow.com/questions/20241857/android-intent-cannot-resolve-constructor
-                Intent layoutSelectIntent = new Intent(v.getContext(), ThreePlayerActivity.class );
+                Intent layoutSelectIntent = new Intent(v.getContext(), GameActivity.class );
+                layoutSelectIntent.putExtra("playerCount", 3);
                 startActivity(layoutSelectIntent);
 
             }
@@ -63,7 +64,7 @@ public class PlayerSelectActivity extends AppCompatActivity {
             public void onClick(View v) {
                 // on click, send user to two player layout selection screen and store choice
                 Intent layoutSelectIntent = new Intent(v.getContext(), LayoutSelectActivity.class );
-                layoutSelectIntent.putExtra("playerSelectCount", 4);
+                layoutSelectIntent.putExtra("playerCount", 4);
                 startActivity(layoutSelectIntent);
 
             }
