+++ app/src/main/res/layout/activity_four_player1.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".FourPlayerActivity">
+    tools:context=".GameActivity">
 
     <android.support.constraint.Guideline
         android:id="@+id/hG2"
