+++ app/src/main/res/layout/activity_two_player2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".TwoPlayerActivity">
+    tools:context=".GameActivity">
 
     <LinearLayout
         android:id="@+id/p1Container"
